Improvements to the Obama command, including optimisations
This fixes an error where the bot would break when the server is down.
This also includes optimisations which reduces the number of requests
made to the server (from nearly infinity to maximum 3). Also included
are the removal of the bot response error (in preference to the discord.
HTTPException error), nicer error messages, and a check which checks if
http://talkobamato.me is down (and caches)