Fixes numerous bugs with the raw command
These include `AttributeError: 'str' object has no attribute 'embeds'`
in multiple places as was using a message content (str) instead of a
message. Also, `TypeError: expected string or bytes-like object` when
it wasn't a message.

Also creates a FakeMessage class to handle when only text (not message)
is received, this contains the text (FakeMessage.content) and an empty
embed list.

This also moves the clean_content function out into a static method
outside of a class.